Тест по power

+++ 03.AlgebraicAlgorithm/src/main/power/IterationPow.java
@@ -2,8 +2,8 @@ package main.power;
 
 public class IterationPow {
 
-    public int pow(int a, int n) {
-        int result = 1;
+    public double pow(double a, int n) {
+        double result = 1;
         for (; n > 0; n--) {
             result *= a;
         }
